Clean up error handling in the decrypting reader
Straighten out the logic by using match and not if let. More importantly,
handle the case where the secret stream's `pull()` operation fails by
returning Poll::Pending, which has the effect of making appear that there
is no data. This seems like a better thing to do than panicking or closing
the connection, but more testing is needed.

The goal is to get rid of all the panicking actions, but this only gets
rid of one (but it is the only one I've seen so far).