Convert to a reliable (and simpler) transport scheme.
This replaces the copy future that moved data between split TcpStream
parts that had custom AsyncRead-implementing wrappers. That technique
wasn't reliable because it assumed that a send of N bytes always
corresponded to a read of N bytes on the remote system, something that
TCP does not guarantee.

Instead, the data transfer is done by a pair of async fns that work in
concert to pass fixed sized (with padding if needed) encrypted message
blocks.

In the README, add a section on the proxy technique and revise the
acknowledgements section. Add a brief note on the use of `RUST_LOG`
for controlling diagnostics.